requirements.txt

web3==6.15.1
eth-account==0.10.0
requests==2.31.0
python-dotenv==1.0.0
eth-account
python-dotenv
requests
web3
flask
flask
pytest
numpy
pandas
scikit-learn
urllib3
